manifest: add a category tag to the intent filter
Not sure if that's the root cause, but it seems that Sony Xperia
phones were not able to use the app to share photos without that
change.

Well, either that, or something else happened to the user who reported
it and reinstalling the new version fixed the problem regardless of
the manifest change it contained.